source: hoist shared forge helpers; drop ForgejoLikeKind
